+++ detector/src/main/java/nl/littlerobots/rxlint/RxJavaSubscriberCheck.java
@@ -9,11 +9,13 @@ public class RxJavaSubscriberCheck implements SubscribeDetector.SubscriberCheck 
     public boolean isMissingOnError(PsiMethod method) {
         PsiClass clz = method.getContainingClass();
         if ("rx.Observable".equals(clz.getQualifiedName()) || "rx.Single".equals(clz.getQualifiedName())) {
-            return method.getParameterList().getParametersCount() == 1 &&
-                    TypeConversionUtil.erasure(method.getParameterList().getParameters()[0].getType()).equalsToText("rx.functions.Action1");
+            return method.getParameterList().getParametersCount() == 0 ||
+                    (method.getParameterList().getParametersCount() == 1 &&
+                            TypeConversionUtil.erasure(method.getParameterList().getParameters()[0].getType()).equalsToText("rx.functions.Action1"));
         } else if ("rx.Completable".equals(clz.getQualifiedName())) {
             // only a completion callback
-            return method.getParameterList().getParametersCount() == 1 && TypeConversionUtil.erasure(method.getParameterList().getParameters()[0].getType()).equalsToText("rx.functions.Action0");
+            return method.getParameterList().getParametersCount() == 0 ||
+                    (method.getParameterList().getParametersCount() == 1 && TypeConversionUtil.erasure(method.getParameterList().getParameters()[0].getType()).equalsToText("rx.functions.Action0"));
         }
         return false;
     }