+++ detector/src/test/java/nl/littlerobots/rxlint/SubscribeDetectorTest.java
@@ -36,36 +36,26 @@ public class SubscribeDetectorTest extends LintDetectorTest {
         return Collections.singletonList(SubscribeDetector.ISSUE);
     }
 
-    public void testSubscribeCheck() throws Exception {
-        String result = lintProject(TestFiles.copy("rxjava-1.1.2.jar", "libs/rxjava.jar", this),
-                TestFiles.copy("rxjava-2.0.5.jar", "libs/rxjava2.jar", this),
-                TestFiles.copy("reactive-streams-1.0.0.final.jar", "libs/reactive-streams-1.0.0.final.jar", this),
+    public void testSubscribeCheckRx1() throws Exception {
+        String result = lintProject(TestFiles.copy("rxjava-1.3.0.jar", "libs/rxjava.jar", this),
                 TestFiles.copy("testjavalib.jar", "libs/testjavalib.jar", this),
                 TestFiles.java("package nl.littlerobots.testproject;\n" +
                         "\n" +
                         "import java.util.concurrent.Callable;\n" +
                         "\n" +
-                        "import io.reactivex.disposables.Disposable;\n" +
-                        "import io.reactivex.functions.Action;\n" +
-                        "import io.reactivex.functions.BiConsumer;\n" +
-                        "import io.reactivex.functions.Consumer;\n" +
-                        "import io.reactivex.observers.DisposableCompletableObserver;\n" +
-                        "import io.reactivex.observers.DisposableMaybeObserver;\n" +
-                        "import io.reactivex.observers.DisposableObserver;\n" +
-                        "import io.reactivex.observers.DisposableSingleObserver;\n" +
-                        "import io.reactivex.subscribers.DisposableSubscriber;\n" +
                         "import nl.littlerobots.test.MySubscriber;\n" +
                         "import rx.Completable;\n" +
                         "import rx.Observable;\n" +
                         "import rx.Observer;\n" +
                         "import rx.Single;\n" +
+                        "import rx.SingleSubscriber;\n" +
                         "import rx.Subscriber;\n" +
                         "import rx.Subscription;\n" +
                         "import rx.functions.Action0;\n" +
                         "import rx.functions.Action1;\n" +
                         "\n" +
-                        "public class SubscriberTest {\n" +
-                        "    public void subscribeWithError() {\n" +
+                        "public class SubscriberTestRx1 {\n" +
+                        "    public void subscribeObservableWithError() {\n" +
                         "        Subscription s = Observable.<String>just(null).subscribe(new Action1<String>() {\n" +
                         "            @Override\n" +
                         "            public void call(String s) {\n" +
@@ -79,7 +69,11 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "        });\n" +
                         "    }\n" +
                         "\n" +
-                        "    public void subscribeWithoutError() {\n" +
+                        "    public void subscribeObservableNoArguments() {\n" +
+                        "        Observable.just(\"Test\").subscribe();\n" +
+                        "    }\n" +
+                        "\n" +
+                        "    public void subscribeObservableWithoutError() {\n" +
                         "        Subscription s = Observable.<String>just(null).subscribe(new Action1<String>() {\n" +
                         "            @Override\n" +
                         "            public void call(String s) {\n" +
@@ -88,7 +82,7 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "        });\n" +
                         "    }\n" +
                         "\n" +
-                        "    public void subscribeWithSubscriber() {\n" +
+                        "    public void subscribeObservableWithSubscriber() {\n" +
                         "        Subscription s = Observable.<String>just(null).subscribe(new Subscriber<String>() {\n" +
                         "            @Override\n" +
                         "            public void onCompleted() {\n" +
@@ -107,7 +101,7 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "        });\n" +
                         "    }\n" +
                         "\n" +
-                        "    public void subscribeWithObserver() {\n" +
+                        "    public void subscribeObservableWithObserver() {\n" +
                         "        Subscription s = Observable.<String>just(null).subscribe(new Observer<String>() {\n" +
                         "            @Override\n" +
                         "            public void onCompleted() {\n" +
@@ -162,6 +156,43 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "        });\n" +
                         "    }\n" +
                         "\n" +
+                        "    public void singleNoArguments() {\n" +
+                        "        Single.just(null).subscribe();\n" +
+                        "    }\n" +
+                        "\n" +
+                        "    public void singleWithSingleSubscriber() {\n" +
+                        "        Single.just(\"test\").subscribe(new SingleSubscriber<String>() {\n" +
+                        "            @Override\n" +
+                        "            public void onSuccess(String value) {\n" +
+                        "\n" +
+                        "            }\n" +
+                        "\n" +
+                        "            @Override\n" +
+                        "            public void onError(Throwable error) {\n" +
+                        "\n" +
+                        "            }\n" +
+                        "        });\n" +
+                        "    }\n" +
+                        "\n" +
+                        "    public void singleWithSubcriber() {\n" +
+                        "        Single.just(\"test\").subscribe(new Subscriber<String>() {\n" +
+                        "            @Override\n" +
+                        "            public void onCompleted() {\n" +
+                        "\n" +
+                        "            }\n" +
+                        "\n" +
+                        "            @Override\n" +
+                        "            public void onError(Throwable e) {\n" +
+                        "\n" +
+                        "            }\n" +
+                        "\n" +
+                        "            @Override\n" +
+                        "            public void onNext(String s) {\n" +
+                        "\n" +
+                        "            }\n" +
+                        "        });\n" +
+                        "    }\n" +
+                        "\n" +
                         "    public void subscribeWithJavaLibSubscriber() {\n" +
                         "        Subscription s = Observable.just(null).subscribe(new MySubscriber<>());\n" +
                         "    }\n" +
@@ -172,15 +203,15 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "            public Object call() throws Exception {\n" +
                         "                return null;\n" +
                         "            }\n" +
-                        "        }).subscribe(new Action1<Throwable>() {\n" +
-                        "            @Override\n" +
-                        "            public void call(Throwable error) {\n" +
-                        "\n" +
-                        "            }\n" +
-                        "        }, new Action0() {\n" +
+                        "        }).subscribe(new Action0() {\n" +
                         "            @Override\n" +
                         "            public void call() {\n" +
                         "\n" +
+                        "            }\n" +
+                        "        }, new Action1<Throwable>() {\n" +
+                        "            @Override\n" +
+                        "            public void call(Throwable throwable) {\n" +
+                        "                \n" +
                         "            }\n" +
                         "        });\n" +
                         "    }\n" +
@@ -223,8 +254,55 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "        });\n" +
                         "    }\n" +
                         "\n" +
-                        "    // Rx2 types\n" +
+                        "    public void completableWithOutSubscriber() {\n" +
+                        "        Completable.fromAction(new Action0() {\n" +
+                        "            @Override\n" +
+                        "            public void call() {\n" +
                         "\n" +
+                        "            }\n" +
+                        "        }).subscribe();\n" +
+                        "    }\n" +
+                        "}\n"));
+
+        assertEquals("src/nl/littlerobots/testproject/SubscriberTestRx1.java:32: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "        Observable.just(\"Test\").subscribe();\n" +
+                "        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx1.java:36: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "        Subscription s = Observable.<String>just(null).subscribe(new Action1<String>() {\n" +
+                "                         ^\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx1.java:110: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "        Subscription s = Single.<String>just(null).subscribe(new Action1<String>() {\n" +
+                "                         ^\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx1.java:119: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "        Single.just(null).subscribe();\n" +
+                "        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx1.java:179: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "        Completable.fromCallable(new Callable<Object>() {\n" +
+                "        ^\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx1.java:217: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "        Completable.fromAction(new Action0() {\n" +
+                "        ^\n" +
+                "6 errors, 0 warnings\n", result);
+    }
+
+
+    public void testSubscribeCheckRx2() throws Exception {
+        String result = lintProject(TestFiles.copy("rxjava-2.0.5.jar", "libs/rxjava2.jar", this),
+                TestFiles.copy("reactive-streams-1.0.0.final.jar", "libs/reactive-streams-1.0.0.final.jar", this),
+                TestFiles.copy("testjavalib.jar", "libs/testjavalib.jar", this),
+                TestFiles.java("package nl.littlerobots.testproject;\n" +
+                        "\n" +
+                        "import io.reactivex.disposables.Disposable;\n" +
+                        "import io.reactivex.functions.Action;\n" +
+                        "import io.reactivex.functions.BiConsumer;\n" +
+                        "import io.reactivex.functions.Consumer;\n" +
+                        "import io.reactivex.observers.DisposableCompletableObserver;\n" +
+                        "import io.reactivex.observers.DisposableMaybeObserver;\n" +
+                        "import io.reactivex.observers.DisposableObserver;\n" +
+                        "import io.reactivex.observers.DisposableSingleObserver;\n" +
+                        "import io.reactivex.subscribers.DisposableSubscriber;\n" +
+                        "\n" +
+                        "public class SubscriberTestRx2 {\n" +
                         "    public void rx2ObservableWithSubscriber() {\n" +
                         "        io.reactivex.Observable.just(\"test\").subscribeWith(new DisposableObserver<String>() {\n" +
                         "            @Override\n" +
@@ -265,6 +343,10 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "\n" +
                         "            }\n" +
                         "        });\n" +
+                        "    }\n" +
+                        "\n" +
+                        "    public void rx2ObservableNoArguments() {\n" +
+                        "        io.reactivex.Observable.just(\"test\").subscribe();\n" +
                         "    }\n" +
                         "\n" +
                         "    public void rx2FlowableWithSubscriber() {\n" +
@@ -309,6 +391,10 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "        });\n" +
                         "    }\n" +
                         "\n" +
+                        "    public void rx2FlowableNoArguments() {\n" +
+                        "        io.reactivex.Flowable.just(\"test\").subscribe();\n" +
+                        "    }\n" +
+                        "\n" +
                         "    public void rx2CompletableWithSubscriber() {\n" +
                         "        io.reactivex.Completable.complete().subscribeWith(new DisposableCompletableObserver() {\n" +
                         "            @Override\n" +
@@ -345,6 +431,11 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "        });\n" +
                         "    }\n" +
                         "\n" +
+                        "    public void rx2CompletableNoArguments() {\n" +
+                        "        io.reactivex.Completable.complete().subscribe();\n" +
+                        "    }\n" +
+                        "\n" +
+                        "\n" +
                         "    public void rx2SingleWithSubscriber() {\n" +
                         "        io.reactivex.Single.just(\"test\").subscribeWith(new DisposableSingleObserver<String>() {\n" +
                         "            @Override\n" +
@@ -358,6 +449,7 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "            }\n" +
                         "        });\n" +
                         "    }\n" +
+                        "\n" +
                         "\n" +
                         "    public void rx2SingleWithOnError() {\n" +
                         "        io.reactivex.Single.just(\"test\").subscribe(new Consumer<String>() {\n" +
@@ -380,6 +472,10 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "\n" +
                         "            }\n" +
                         "        });\n" +
+                        "    }\n" +
+                        "\n" +
+                        "    public void rx2SingleNoArguments() {\n" +
+                        "        io.reactivex.Single.just(\"test\").subscribe();\n" +
                         "    }\n" +
                         "\n" +
                         "    public void rx2MaybeWithSubscriber() {\n" +
@@ -424,6 +520,11 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "        });\n" +
                         "    }\n" +
                         "\n" +
+                        "    public void rx2MaybeNoArguments() {\n" +
+                        "        io.reactivex.Maybe.just(\"test\").subscribe();\n" +
+                        "    }\n" +
+                        "\n" +
+                        "\n" +
                         "    public void rx2WithObserver() {\n" +
                         "        io.reactivex.Observable.just(\"test\").subscribe(new io.reactivex.Observer<String>() {\n" +
                         "            @Override\n" +
@@ -450,15 +551,18 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "\n" +
                         "    public void rx2WithSingleObserver() {\n" +
                         "        io.reactivex.Single.just(1).subscribe(new io.reactivex.SingleObserver<Integer>() {\n" +
-                        "            @Override public void onSubscribe(Disposable d) {\n" +
+                        "            @Override\n" +
+                        "            public void onSubscribe(Disposable d) {\n" +
                         "\n" +
                         "            }\n" +
                         "\n" +
-                        "            @Override public void onSuccess(Integer integer) {\n" +
+                        "            @Override\n" +
+                        "            public void onSuccess(Integer integer) {\n" +
                         "\n" +
                         "            }\n" +
                         "\n" +
-                        "            @Override public void onError(Throwable e) {\n" +
+                        "            @Override\n" +
+                        "            public void onError(Throwable e) {\n" +
                         "\n" +
                         "            }\n" +
                         "        });\n" +
@@ -466,19 +570,23 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "\n" +
                         "    public void rx2WithMaybeObserver() {\n" +
                         "        io.reactivex.Maybe.just(1).subscribe(new io.reactivex.MaybeObserver<Integer>() {\n" +
-                        "            @Override public void onSubscribe(Disposable d) {\n" +
+                        "            @Override\n" +
+                        "            public void onSubscribe(Disposable d) {\n" +
                         "\n" +
                         "            }\n" +
                         "\n" +
-                        "            @Override public void onSuccess(Integer integer) {\n" +
+                        "            @Override\n" +
+                        "            public void onSuccess(Integer integer) {\n" +
                         "\n" +
                         "            }\n" +
                         "\n" +
-                        "            @Override public void onError(Throwable e) {\n" +
+                        "            @Override\n" +
+                        "            public void onError(Throwable e) {\n" +
                         "\n" +
                         "            }\n" +
                         "\n" +
-                        "            @Override public void onComplete() {\n" +
+                        "            @Override\n" +
+                        "            public void onComplete() {\n" +
                         "\n" +
                         "            }\n" +
                         "        });\n" +
@@ -486,15 +594,18 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "\n" +
                         "    public void rx2WithCompletableObserver() {\n" +
                         "        io.reactivex.Completable.complete().subscribe(new io.reactivex.CompletableObserver() {\n" +
-                        "            @Override public void onSubscribe(Disposable d) {\n" +
+                        "            @Override\n" +
+                        "            public void onSubscribe(Disposable d) {\n" +
                         "\n" +
                         "            }\n" +
                         "\n" +
-                        "            @Override public void onComplete() {\n" +
+                        "            @Override\n" +
+                        "            public void onComplete() {\n" +
                         "\n" +
                         "            }\n" +
                         "\n" +
-                        "            @Override public void onError(Throwable e) {\n" +
+                        "            @Override\n" +
+                        "            public void onError(Throwable e) {\n" +
                         "\n" +
                         "            }\n" +
                         "        });\n" +
@@ -502,38 +613,46 @@ public class SubscribeDetectorTest extends LintDetectorTest {
                         "\n" +
                         "    public void rx2WithSingleBiConsumer() {\n" +
                         "        io.reactivex.Single.just(1).subscribe(new BiConsumer<Integer, Throwable>() {\n" +
-                        "            @Override public void accept(Integer integer, Throwable throwable) throws Exception {\n" +
+                        "            @Override\n" +
+                        "            public void accept(Integer integer, Throwable throwable) throws Exception {\n" +
                         "\n" +
                         "            }\n" +
                         "        });\n" +
                         "    }\n" +
+                        "\n" +
                         "}\n"));
 
-        assertEquals("src/nl/littlerobots/testproject/SubscriberTest.java:40: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
-                "        Subscription s = Observable.<String>just(null).subscribe(new Action1<String>() {\n" +
-                "                         ^\n" +
-                "src/nl/littlerobots/testproject/SubscriberTest.java:114: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
-                "        Subscription s = Single.<String>just(null).subscribe(new Action1<String>() {\n" +
-                "                         ^\n" +
-                "src/nl/littlerobots/testproject/SubscriberTest.java:146: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
-                "        Completable.fromCallable(new Callable<Object>() {\n" +
-                "        ^\n" +
-                "src/nl/littlerobots/testproject/SubscriberTest.java:219: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+        assertEquals("src/nl/littlerobots/testproject/SubscriberTestRx2.java:48: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
                 "        io.reactivex.Observable.just(\"test\").subscribe(new Consumer<String>() {\n" +
                 "        ^\n" +
-                "src/nl/littlerobots/testproject/SubscriberTest.java:261: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx2.java:57: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "        io.reactivex.Observable.just(\"test\").subscribe();\n" +
+                "        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx2.java:94: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
                 "        io.reactivex.Flowable.just(\"Test\").subscribe(new Consumer<String>() {\n" +
                 "        ^\n" +
-                "src/nl/littlerobots/testproject/SubscriberTest.java:298: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx2.java:103: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "        io.reactivex.Flowable.just(\"test\").subscribe();\n" +
+                "        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx2.java:135: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
                 "        io.reactivex.Completable.complete().subscribe(new Action() {\n" +
                 "        ^\n" +
-                "src/nl/littlerobots/testproject/SubscriberTest.java:334: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx2.java:143: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "        io.reactivex.Completable.complete().subscribe();\n" +
+                "        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx2.java:177: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
                 "        io.reactivex.Single.just(\"test\").subscribe(new Consumer<String>() {\n" +
                 "        ^\n" +
-                "src/nl/littlerobots/testproject/SubscriberTest.java:376: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx2.java:186: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "        io.reactivex.Single.just(\"test\").subscribe();\n" +
+                "        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n" +
+                "src/nl/littlerobots/testproject/SubscriberTestRx2.java:223: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
                 "        io.reactivex.Maybe.just(\"test\").subscribe(new Consumer<String>() {\n" +
                 "        ^\n" +
-                "8 errors, 0 warnings\n", result);
+                "src/nl/littlerobots/testproject/SubscriberTestRx2.java:232: Error: Subscriber is missing onError [RxSubscribeOnError]\n" +
+                "        io.reactivex.Maybe.just(\"test\").subscribe();\n" +
+                "        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n" +
+                "10 errors, 0 warnings\n", result);
     }
 
     @Override