+++ testproject/src/main/java/nl/littlerobots/testproject/SubscriberTestRx2.java
@@ -1,22 +1,4 @@
-/*
- *    Copyright 2016 Little Robots
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
 package nl.littlerobots.testproject;
-
-import java.util.concurrent.Callable;
 
 import io.reactivex.disposables.Disposable;
 import io.reactivex.functions.Action;
@@ -27,177 +9,8 @@ import io.reactivex.observers.DisposableMaybeObserver;
 import io.reactivex.observers.DisposableObserver;
 import io.reactivex.observers.DisposableSingleObserver;
 import io.reactivex.subscribers.DisposableSubscriber;
-import nl.littlerobots.test.MySubscriber;
-import rx.Completable;
-import rx.Observable;
-import rx.Observer;
-import rx.Single;
-import rx.Subscriber;
-import rx.Subscription;
-import rx.functions.Action0;
-import rx.functions.Action1;
 
-public class SubscriberTest {
-    public void subscribeWithError() {
-        Subscription s = Observable.<String>just(null).subscribe(new Action1<String>() {
-            @Override
-            public void call(String s) {
-
-            }
-        }, new Action1<Throwable>() {
-            @Override
-            public void call(Throwable throwable) {
-
-            }
-        });
-    }
-
-    public void subscribeWithoutError() {
-        Subscription s = Observable.<String>just(null).subscribe(new Action1<String>() {
-            @Override
-            public void call(String s) {
-
-            }
-        });
-    }
-
-    public void subscribeWithSubscriber() {
-        Subscription s = Observable.<String>just(null).subscribe(new Subscriber<String>() {
-            @Override
-            public void onCompleted() {
-
-            }
-
-            @Override
-            public void onError(Throwable e) {
-
-            }
-
-            @Override
-            public void onNext(String s) {
-
-            }
-        });
-    }
-
-    public void subscribeWithObserver() {
-        Subscription s = Observable.<String>just(null).subscribe(new Observer<String>() {
-            @Override
-            public void onCompleted() {
-
-            }
-
-            @Override
-            public void onError(Throwable e) {
-
-            }
-
-            @Override
-            public void onNext(String s) {
-
-            }
-        });
-    }
-
-    public void testCallingRandomSubcribe() {
-        subscribe(new Action1<String>() {
-            @Override
-            public void call(String s) {
-
-            }
-        });
-    }
-
-    private void subscribe(Action1<String> test) {
-
-    }
-
-    public void singleWithError() {
-        Subscription s = Single.<String>just(null).subscribe(new Action1<String>() {
-            @Override
-            public void call(String s) {
-
-            }
-        }, new Action1<Throwable>() {
-            @Override
-            public void call(Throwable throwable) {
-
-            }
-        });
-    }
-
-    public void singleWithoutError() {
-        Subscription s = Single.<String>just(null).subscribe(new Action1<String>() {
-            @Override
-            public void call(String s) {
-
-            }
-        });
-    }
-
-    public void subscribeWithJavaLibSubscriber() {
-        Subscription s = Observable.just(null).subscribe(new MySubscriber<>());
-    }
-
-    public void completableWithError() {
-        Completable.fromCallable(new Callable<Object>() {
-            @Override
-            public Object call() throws Exception {
-                return null;
-            }
-        }).subscribe(new Action1<Throwable>() {
-            @Override
-            public void call(Throwable error) {
-
-            }
-        }, new Action0() {
-            @Override
-            public void call() {
-
-            }
-        });
-    }
-
-    public void completableWithoutError() {
-        Completable.fromCallable(new Callable<Object>() {
-            @Override
-            public Object call() throws Exception {
-                return null;
-            }
-        }).subscribe(new Action0() {
-            @Override
-            public void call() {
-
-            }
-        });
-    }
-
-    public void completableWithSubscriber() {
-        Completable.fromCallable(new Callable<Object>() {
-            @Override
-            public Object call() throws Exception {
-                return null;
-            }
-        }).subscribe(new Subscriber<Object>() {
-            @Override
-            public void onCompleted() {
-
-            }
-
-            @Override
-            public void onError(Throwable e) {
-
-            }
-
-            @Override
-            public void onNext(Object o) {
-
-            }
-        });
-    }
-
-    // Rx2 types
-
+public class SubscriberTestRx2 {
     public void rx2ObservableWithSubscriber() {
         io.reactivex.Observable.just("test").subscribeWith(new DisposableObserver<String>() {
             @Override
@@ -238,6 +51,10 @@ public class SubscriberTest {
 
             }
         });
+    }
+
+    public void rx2ObservableNoArguments() {
+        io.reactivex.Observable.just("test").subscribe();
     }
 
     public void rx2FlowableWithSubscriber() {
@@ -282,6 +99,10 @@ public class SubscriberTest {
         });
     }
 
+    public void rx2FlowableNoArguments() {
+        io.reactivex.Flowable.just("test").subscribe();
+    }
+
     public void rx2CompletableWithSubscriber() {
         io.reactivex.Completable.complete().subscribeWith(new DisposableCompletableObserver() {
             @Override
@@ -318,6 +139,11 @@ public class SubscriberTest {
         });
     }
 
+    public void rx2CompletableNoArguments() {
+        io.reactivex.Completable.complete().subscribe();
+    }
+
+
     public void rx2SingleWithSubscriber() {
         io.reactivex.Single.just("test").subscribeWith(new DisposableSingleObserver<String>() {
             @Override
@@ -331,6 +157,7 @@ public class SubscriberTest {
             }
         });
     }
+
 
     public void rx2SingleWithOnError() {
         io.reactivex.Single.just("test").subscribe(new Consumer<String>() {
@@ -353,6 +180,10 @@ public class SubscriberTest {
 
             }
         });
+    }
+
+    public void rx2SingleNoArguments() {
+        io.reactivex.Single.just("test").subscribe();
     }
 
     public void rx2MaybeWithSubscriber() {
@@ -396,6 +227,11 @@ public class SubscriberTest {
             }
         });
     }
+
+    public void rx2MaybeNoArguments() {
+        io.reactivex.Maybe.just("test").subscribe();
+    }
+
 
     public void rx2WithObserver() {
         io.reactivex.Observable.just("test").subscribe(new io.reactivex.Observer<String>() {
@@ -491,4 +327,5 @@ public class SubscriberTest {
             }
         });
     }
+
 }